+++ exe/arduino_ci_remote.rb
@@ -55,7 +55,16 @@ end
 
 # Make a nice status for something that kills the script immediately on failure
 def assure(message, &block)
-  perform_action(message, false, "This may indicate a problem with ArduinoCI!", true, &block)
+  perform_action(message, false, "This may indicate a problem with ArduinoCI, or your configuration", true, &block)
+end
+
+# Assure that a platform exists and return its definition
+def assured_platform(purpose, name, config)
+  platform_definition = config.platform_definition(name)
+  assure("Requested #{purpose} platform '#{name}' is defined in 'platforms' YML") do
+    !platform_definition.nil?
+  end
+  platform_definition
 end
 
 # initialize command and config
@@ -85,10 +94,11 @@ end
 # while we're doing that, get the aux libraries as well
 all_platforms = {}
 aux_libraries = Set.new(config.aux_libraries_for_unittest + config.aux_libraries_for_build)
-config.platforms_to_unittest.each { |p| all_platforms[p] = config.platform_definition(p) }
+# while collecting the platforms, ensure they're defined
+config.platforms_to_unittest.each { |p| all_platforms[p] = assured_platform("unittest", p, config) }
 library_examples.each do |path|
   ovr_config = config.from_example(path)
-  ovr_config.platforms_to_build.each { |p| all_platforms[p] = config.platform_definition(p) }
+  ovr_config.platforms_to_build.each { |p| all_platforms[p] = assured_platform("library example", p, config) }
   aux_libraries.merge(ovr_config.aux_libraries_for_build)
 end